+++ invenio_app_rdm/records_ui/views/deposits.py
@@ -37,7 +37,11 @@ from invenio_vocabularies.records.models import VocabularyScheme
 from marshmallow_utils.fields.babel import gettext_from_dict
 from sqlalchemy.orm import load_only
 
-from ..utils import evaluate_record_deletion, set_default_value
+from ..utils import (
+    evaluate_file_modification,
+    evaluate_record_deletion,
+    set_default_value,
+)
 from .decorators import (
     no_cache_response,
     pass_draft,
@@ -554,8 +558,12 @@ def deposit_edit(pid_value, draft=None, draft_files=None, files_locked=True):
         record_deletion = evaluate_record_deletion(
             published_record_result._record, g.identity
         )
+        file_modification = evaluate_file_modification(
+            published_record_result._record, g.identity
+        )
     else:
         record_deletion = {}
+        file_modification = {}
 
     community_ui = None
     community_theme = None
@@ -631,6 +639,7 @@ def deposit_edit(pid_value, draft=None, draft_files=None, files_locked=True):
             ]
         ),
         record_deletion=record_deletion,
+        file_modification=file_modification,
     )