+++ invenio_app_rdm/records_ui/views/records.py
@@ -15,6 +15,7 @@ from pathlib import Path
 
 from flask import abort, current_app, g, redirect, render_template, request, url_for
 from flask_login import current_user
+from flask_principal import AnonymousIdentity
 from invenio_base.utils import obj_or_import_string
 from invenio_communities.communities.resources.serializer import (
     UICommunityJSONSerializer,
@@ -28,7 +29,10 @@ from invenio_rdm_records.proxies import current_rdm_records
 from invenio_rdm_records.records.systemfields.access.access_settings import (
     AccessSettings,
 )
+from invenio_rdm_records.requests import CommunityInclusion, CommunitySubmission
 from invenio_rdm_records.resources.serializers import UIJSONSerializer
+from invenio_requests.proxies import current_requests_service
+from invenio_search.api import dsl
 from invenio_stats.proxies import current_stats
 from invenio_users_resources.proxies import current_user_resources
 from marshmallow import ValidationError
@@ -45,7 +49,9 @@ from .decorators import (
     pass_file_item,
     pass_file_metadata,
     pass_include_deleted,
+    pass_is_iframe,
     pass_is_preview,
+    pass_preview_file,
     pass_record_files,
     pass_record_from_pid,
     pass_record_latest,
@@ -99,6 +105,43 @@ def get_record_community(record):
         return None, None
 
 
+def get_record_requests(record, identity):
+    """Return all requests that concern this record.
+
+    Output: {<Community-UUID>: <Request-UUID>}
+    """
+    can_review = current_rdm_records.records_service.check_permission(
+        identity, "review", record=record._record
+    )
+    if not can_review:
+        return {}
+
+    if type(identity) is AnonymousIdentity:
+        return {}  # secret link users do not have permissions to search requests
+
+    record_requests = current_requests_service.search(
+        identity,
+        extra_filter=dsl.Q(
+            "bool",
+            must=[
+                dsl.Q("term", **{"topic.record": record["id"]}),
+                dsl.Q(
+                    "terms",
+                    **{
+                        "type": [
+                            CommunityInclusion.type_id,
+                            CommunitySubmission.type_id,
+                        ]
+                    },
+                ),
+            ],
+        ),
+        params={"sort": "oldest"},
+    )
+
+    return {r["receiver"]["community"]: r["id"] for r in record_requests}
+
+
 class PreviewFile:
     """Preview file implementation for InvenioRDM.
 
@@ -106,7 +149,9 @@ class PreviewFile:
     `invenio_previewer.api.PreviewFile`.
     """
 
-    def __init__(self, file_item, record_pid_value, record=None, url=None):
+    def __init__(
+        self, file_item, record_pid_value, record=None, url=None, is_iframe=False
+    ):
         """Create a new PreviewFile."""
         self.file = file_item
         self.data = file_item.data
@@ -119,6 +164,7 @@ class PreviewFile:
             pid_value=record_pid_value,
             filename=self.filename,
         )
+        self.is_iframe = is_iframe
 
     def is_local(self):
         """Check if file is local."""
@@ -144,12 +190,19 @@ class PreviewFile:
 
 @pass_is_preview
 @pass_include_deleted
+@pass_preview_file
 @pass_record_or_draft(expand=True)
 @pass_record_files
 @pass_record_media_files
 @add_signposting_landing_page
 def record_detail(
-    pid_value, record, files, media_files, is_preview=False, include_deleted=False
+    pid_value,
+    record,
+    files,
+    media_files,
+    is_preview=False,
+    include_deleted=False,
+    preview_file=None,
 ):
     """Record detail page (aka landing page)."""
     files_dict = None if files is None else files.to_dict()
@@ -160,6 +213,9 @@ def record_detail(
         record._record.parent["access"]["settings"] = AccessSettings({}).dump()
 
     record_ui = UIJSONSerializer().dump_obj(record.to_dict())
+
+    if preview_file:
+        record_ui["files"]["default_preview"] = preview_file
 
     record_deletion = evaluate_record_deletion(record._record, g.identity)
 
@@ -238,6 +294,8 @@ def record_detail(
     )
     theme = resolved_community_ui.get("theme", {}) if resolved_community else None
 
+    record_requests = get_record_requests(record, g.identity)
+
     return render_community_theme_template(
         current_app.config.get("APP_RDM_RECORD_LANDING_PAGE_TEMPLATE"),
         theme=theme,
@@ -267,6 +325,7 @@ def record_detail(
         is_draft=is_draft,
         community=resolved_community,
         community_ui=resolved_community_ui,
+        record_requests=record_requests,
         external_resources=get_external_resources(record),
         user_avatar=avatar,
         record_deletion=record_deletion,
@@ -303,6 +362,7 @@ def record_export(
 @pass_include_deleted
 @pass_record_or_draft(expand=False)
 @pass_file_metadata
+@pass_is_iframe
 def record_file_preview(
     pid_value,
     record=None,
@@ -310,6 +370,7 @@ def record_file_preview(
     file_metadata=None,
     is_preview=False,
     include_deleted=False,
+    is_iframe=False,
     **kwargs,
 ):
     """Render a preview of the specified file."""
@@ -321,7 +382,7 @@ def record_file_preview(
     )
 
     # Find a suitable previewer
-    fileobj = PreviewFile(file_metadata, pid_value, record, url)
+    fileobj = PreviewFile(file_metadata, pid_value, record, url, is_iframe)
     # Try to see if specific previewer preference is set for the file
     file_previewer = (file_metadata.data.get("metadata") or {}).get("previewer")
     if file_previewer: