+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/deposit/RDMDepositForm.js
@@ -48,6 +48,7 @@ import { CopyrightsField } from "@js/invenio_rdm_records/src/deposit/fields/Copy
 import { ShareDraftButton } from "./ShareDraftButton";
 import { depositFormSectionsConfig, severityChecksConfig } from "./config";
 import { RecordDeletion } from "../components/RecordDeletion";
+import { FileModificationUntil } from "../components/FileModificationUntil";
 
 export class RDMDepositForm extends Component {
   constructor(props) {
@@ -114,6 +115,7 @@ export class RDMDepositForm extends Component {
       recordRestrictionGracePeriod,
       allowRecordRestriction,
       recordDeletion,
+      fileModification,
       groupsEnabled,
       allowEmptyFiles,
       useUppy,
@@ -138,6 +140,7 @@ export class RDMDepositForm extends Component {
         recordRestrictionGracePeriod={recordRestrictionGracePeriod}
         allowRecordRestriction={allowRecordRestriction}
         recordDeletion={recordDeletion}
+        fileModification={fileModification}
         groupsEnabled={groupsEnabled}
         allowEmptyFiles={allowEmptyFiles}
         customFieldsUI={customFieldsUI}
@@ -177,16 +180,46 @@ export class RDMDepositForm extends Component {
             <Grid className="mt-25">
               <Grid.Column mobile={16} tablet={16} computer={11}>
                 <Overridable
+                  id="InvenioAppRdm.Deposit.Files.before.container"
+                  record={record}
+                  files={files}
+                  permissions={permissions}
+                  preselectedCommunity={preselectedCommunity}
+                  filesLocked={filesLocked}
+                  recordRestrictionGracePeriod={recordRestrictionGracePeriod}
+                  allowRecordRestriction={allowRecordRestriction}
+                  groupsEnabled={groupsEnabled}
+                  allowEmptyFiles={allowEmptyFiles}
+                  customFieldsUI={customFieldsUI}
+                  config={this.config}
+                  vocabularies={this.vocabularies}
+                  noFiles={this.noFiles}
+                  hideCommunitySelection={this.hide_community_selection}
+                />
+                <Overridable
                   id="InvenioAppRdm.Deposit.AccordionFieldFiles.container"
                   record={record}
                   config={this.config}
                   noFiles={this.noFiles}
+                  fileModification={fileModification}
+                  filesLocked={filesLocked}
                 >
                   <AccordionField
                     includesPaths={this.sectionsConfig["files-section"]}
                     severityChecks={this.severityChecks}
                     active
-                    label={i18next.t("Files")}
+                    label={
+                      <>
+                        {i18next.t("Files")}
+                        {record.is_published && (
+                          <FileModificationUntil
+                            fileModification={fileModification}
+                            filesLocked={filesLocked}
+                            record={record}
+                          />
+                        )}
+                      </>
+                    }
                     id="files-section"
                   >
                     {this.noFiles && record.is_published && (
@@ -210,10 +243,28 @@ export class RDMDepositForm extends Component {
                         allowEmptyFiles={allowEmptyFiles}
                         filesLocked={filesLocked}
                         fileUploadConcurrency={config.fileUploadConcurrency}
+                        fileModification={fileModification}
                       />
                     </Overridable>
                   </AccordionField>
                 </Overridable>
+                <Overridable
+                  id="InvenioAppRdm.Deposit.Files.after.container"
+                  record={record}
+                  files={files}
+                  permissions={permissions}
+                  preselectedCommunity={preselectedCommunity}
+                  filesLocked={filesLocked}
+                  recordRestrictionGracePeriod={recordRestrictionGracePeriod}
+                  allowRecordRestriction={allowRecordRestriction}
+                  groupsEnabled={groupsEnabled}
+                  allowEmptyFiles={allowEmptyFiles}
+                  customFieldsUI={customFieldsUI}
+                  config={this.config}
+                  vocabularies={this.vocabularies}
+                  noFiles={this.noFiles}
+                  hideCommunitySelection={this.hide_community_selection}
+                />
                 <Overridable
                   id="InvenioAppRdm.Deposit.AccordionFieldBasicInformation.container"
                   config={this.config}
@@ -809,7 +860,8 @@ RDMDepositForm.propTypes = {
   config: PropTypes.object.isRequired,
   recordRestrictionGracePeriod: PropTypes.number.isRequired,
   allowRecordRestriction: PropTypes.bool.isRequired,
-  recordDeletion: PropTypes.object.isRequired,
+  recordDeletion: PropTypes.object,
+  fileModification: PropTypes.object,
   record: PropTypes.object.isRequired,
   preselectedCommunity: PropTypes.object,
   files: PropTypes.object,
@@ -826,4 +878,6 @@ RDMDepositForm.defaultProps = {
   filesLocked: false,
   allowEmptyFiles: true,
   useUppy: false,
+  recordDeletion: {},
+  fileModification: {},
 };