+++ invenio_app_rdm/theme/assets/semantic-ui/js/invenio_app_rdm/landing_page/index.js
@@ -52,10 +52,12 @@ function renderRecordManagement(element) {
 const recordVersionsAppDiv = document.getElementById("recordVersions");
 if (recordVersionsAppDiv) {
   ReactDOM.render(
-    <RecordVersionsList
-      record={JSON.parse(recordVersionsAppDiv.dataset.record)}
-      isPreview={JSON.parse(recordVersionsAppDiv.dataset.preview)}
-    />,
+    <OverridableContext.Provider value={overriddenComponents}>
+      <RecordVersionsList
+        record={JSON.parse(recordVersionsAppDiv.dataset.record)}
+        isPreview={JSON.parse(recordVersionsAppDiv.dataset.preview)}
+      />
+    </OverridableContext.Provider>,
     recordVersionsAppDiv
   );
 }
@@ -101,6 +103,7 @@ if (sidebarCommunitiesManageDiv) {
   );
   const permissions = JSON.parse(sidebarCommunitiesManageDiv.dataset.permissions);
   const record = JSON.parse(sidebarCommunitiesManageDiv.dataset.record);
+  const recordRequests = JSON.parse(sidebarCommunitiesManageDiv.dataset.recordRequests);
   ReactDOM.render(
     <OverridableContext.Provider value={overriddenComponents}>
       <Overridable
@@ -112,6 +115,7 @@ if (sidebarCommunitiesManageDiv) {
         permissions={permissions}
         searchConfig={pendingCommunitiesSearchConfig}
         record={record}
+        recordRequests={recordRequests}
       >
         <CommunitiesManagement
           userCommunitiesMemberships={userCommunitiesMemberships}
@@ -121,6 +125,7 @@ if (sidebarCommunitiesManageDiv) {
           permissions={permissions}
           searchConfig={pendingCommunitiesSearchConfig}
           record={record}
+          recordRequests={recordRequests}
         />
       </Overridable>
     </OverridableContext.Provider>,